Add knowledge of registered formatters for lookup.
The reason that this was put into the `RSpecIntegration` was for
simplicity. All registered formatters in this gem will inherit from
this. This also is a test run for potential changes that will need to be
made to RSpec core to prevent monkey patch it.

Another option was to store these classes in the `EmojiTestLove` module
during the call to `RSpecFormatter`. This was also a valid option. The
only reason this option was not chosen was to test drive a potential
change to RSpec core using the subclass hooks.